requirements.txt
amqp==2.3.2
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
clarifai==2.5.0
configparser==3.5.0
dj-database-url==0.5.0
Django==2.1.4
django-celery-beat==1.4.0
django-paypal==0.5.0
django-timezone-field==3.0
djangorestframework==3.9.0
dnspython==1.16.0
eventlet==0.24.1
future==0.17.1
gevent==1.3.7
googleapis-common-protos==1.5.5
greenlet==0.4.15
grpcio==1.17.0
gunicorn==19.9.0
idna==2.7
jsonschema==2.6.0

monotonic==1.5
numpy==1.15.4
protobuf==3.6.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycparser==2.19
python-crontab==2.3.5
python-dateutil==2.7.5
pytz==2018.7
requests==2.20.1
selenium==3.141.0
six==1.12.0
urllib3==1.24.1
vine==1.1.4
whitenoise==4.1.2
